backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
langchain==0.1.0
openai==1.6.0
anthropic==0.8.0
websockets==12.0
python-multipart==0.0.6
cors==1.0.1
httpx==0.27.2